Fold the dynsize vector tests into codegen_test.go
